clean: movidas las funciones de actualización de códigos a su propio archivo
