+++ actualizador-catedras/db.go
@@ -3,13 +3,10 @@ package main
 import (
 	"context"
 	"errors"
-	"maps"
 	"os"
-	"slices"
 	"time"
 
 	"github.com/charmbracelet/log"
-	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"golang.org/x/sync/errgroup"
 )
@@ -31,187 +28,6 @@ func InitDBPool(logger *log.Logger) error {
 	logger.Debug("establecida conexión con la base de datos")
 
 	return nil
-}
-
-// ActualizarCodigosMaterias sincroniza los códigos de las materias en la
-// base de datos con sus códigos correctos obtenidos del SIU.
-func ActualizarCodigosMaterias(ofertas []oferta) error {
-	logger := log.Default().WithPrefix("🛢️")
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*3)
-	defer cancel()
-
-	var n int
-
-	err := db.QueryRow(ctx, `
-SELECT count(*) FROM materia WHERE codigo LIKE 'COD%'
-		`).Scan(&n)
-
-	if err != nil {
-		log.Error(err)
-		return errors.New("error determinando la cantidad de materias sin actualizar")
-	}
-
-	if n == 0 {
-		logger.Info("no se encontraron materias con códigos sin actualizar")
-		return nil
-	}
-
-	logger.Debugf("encontradas %v materias con códigos sin actualizar", n)
-
-	codigosMaterias, err := getCodigosDeMaterias()
-	if err != nil {
-		logger.Fatal(err)
-		return errors.New("error obteniendo los códigos de las materias de la base de datos")
-	}
-
-	ctx, cancel = context.WithTimeout(context.Background(), time.Second*3)
-	defer cancel()
-
-	tx, err := db.Begin(ctx)
-	if err != nil {
-		logger.Error(err)
-		return errors.New("error iniciando transacción SQL de actualización de códigos")
-	}
-
-	ctx, cancel = context.WithTimeout(context.Background(), time.Second*3)
-	defer cancel()
-
-	_, err = tx.Exec(ctx, `
-CREATE TABLE tmp_codigos_materias (
-	nombre_materia TEXT PRIMARY KEY,
-	codigo_materia_actual TEXT NOT NULL,
-	codigo_materia_siu TEXT NOT NULL
-)
-		`)
-
-	if err != nil {
-		logger.Error(err)
-		return errors.New("error creando tabla temporal de asociación de códigos de materias")
-	}
-
-	rows := make(map[string][]any, len(codigosMaterias))
-
-	for _, o := range ofertas {
-		for _, m := range o.materias {
-			if codActual, ok := codigosMaterias[m.Nombre]; ok {
-				if _, ok := rows[m.Nombre]; !ok {
-					rows[m.Nombre] = []any{m.Nombre, codActual, m.Codigo}
-				}
-			} else {
-				log.Default().WithPrefix("🔎").Warn("materia no está en la base de datos", "codigo", m.Codigo, "nombre", m.Nombre)
-			}
-		}
-	}
-
-	logger.Debugf("obtenidos los códigos de %v materias desde el SIU", len(rows))
-
-	ctx, cancel = context.WithTimeout(context.Background(), time.Second*3)
-	defer cancel()
-
-	_, err = tx.CopyFrom(
-		ctx,
-		pgx.Identifier{"tmp_codigos_materias"},
-		[]string{"nombre_materia", "codigo_materia_actual", "codigo_materia_siu"},
-		pgx.CopyFromRows(slices.Collect(maps.Values(rows))),
-	)
-
-	if err != nil {
-		logger.Error(err)
-		return errors.New("error insertando códigos correctos de mateiras obtenidos del SIU")
-	}
-
-	ctx, cancel = context.WithTimeout(context.Background(), time.Second*3)
-	defer cancel()
-
-	logger.Debug("actualizando los códigos de las materias")
-
-	updateRes, err := tx.Exec(ctx, `
-WITH materias_a_actualizar AS (
-    SELECT m.codigo as codigo_materia_actual, tcm.codigo_materia_siu
-    FROM materia m
-    JOIN tmp_codigos_materias tcm ON lower(unaccent(m.nombre)) = tcm.nombre_materia
-    JOIN plan_materia pm ON m.codigo = pm.codigo_materia
-    JOIN plan p ON pm.codigo_plan = p.codigo
-    WHERE p.esta_vigente = TRUE
-    AND tcm.codigo_materia_actual != tcm.codigo_materia_siu
-)
-UPDATE materia
-SET codigo = ma.codigo_materia_siu
-FROM materias_a_actualizar ma
-WHERE materia.codigo = ma.codigo_materia_actual
-		`)
-
-	if err != nil {
-		logger.Error(err)
-		return errors.New("error actualizando códigos de materias")
-	}
-
-	ctx, cancel = context.WithTimeout(context.Background(), time.Second*3)
-	defer cancel()
-
-	_, err = tx.Exec(ctx, `DROP TABLE tmp_codigos_materias`)
-	if err != nil {
-		logger.Error(err)
-	}
-
-	ctx, cancel = context.WithTimeout(context.Background(), time.Second*3)
-	defer cancel()
-
-	err = tx.Commit(ctx)
-	if err != nil {
-		logger.Error(err)
-		return errors.New("error commiteando transacción SQL de actualización de códigos")
-	}
-
-	logger.Infof("actualizados los códigos de %v materias", updateRes.RowsAffected())
-
-	return nil
-}
-
-// getCodigosDeMaterias retorna un hashmap donde la llave es el nombre de una
-// materia sin diacríticos y en minúscula, y el valor es el código que tiene
-// esa materia en la base de datos.
-func getCodigosDeMaterias() (map[string]string, error) {
-	logger := log.Default().WithPrefix("🛢️")
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*3)
-	defer cancel()
-
-	logger.Debug("obteniendo los códigos de las materias")
-
-	rows, err := db.Query(ctx, `
-SELECT m.codigo, lower(unaccent(m.nombre))
-FROM materia m
-INNER JOIN plan_materia pm
-ON m.codigo = pm.codigo_materia
-INNER JOIN plan p
-ON p.codigo = pm.codigo_plan
-WHERE p.esta_vigente = true
-		`)
-
-	if err != nil {
-		logger.Error(err)
-		return nil, errors.New("error obteniendo los códigos de las materias")
-	}
-
-	codigos := make(map[string]string)
-
-	for rows.Next() {
-		var cod, nombre string
-
-		err := rows.Scan(&cod, &nombre)
-		if err != nil {
-			logger.Error(err, "codigo", cod, "nombre", nombre)
-			return nil, errors.New("error serializando las materias")
-		}
-
-		codigos[nombre] = cod
-	}
-
-	logger.Infof("encontrado los códigos de %v materias", len(codigos))
-
-	return codigos, nil
 }
 
 // getMateriasNoActualizadasEnCuatriActual retorna un hashset con los códigos
